chore: remove N and L from static file and change L type to double

+++ src/main/java/OffLatticeSimulation.java
@@ -5,7 +5,7 @@ import java.util.stream.Collectors;
 
 public class OffLatticeSimulation {
     private List<Particle> particles;
-    private final int L;
+    private final double L;
     private final double eta, dt;
     private int t; // current time
